requirements.txt
numpy
spacy
scikit-learn
sentence-transformers
transformers
fastapi
uvicorn
pydantic
python-dotenv
psycopg2-binary
neo4j
redis
pgvector
openai
httpx
anthropic
faiss-cpu
langchain
langchain-openai
elasticsearch
ragatouille
pytesseract
Pillow
